Crate nic_initialization

Functions that are used in a NIC initialization procedure.

They include allocating memory space for the device’s registers, and initializing its receive and transmit queues.

Constants

NIC_MAPPING_FLAGS
The mapping flags used for pages that the NIC will map.

Functions

allocate_device_register_memory
Allocates memory for the NIC registers

allocate_memory
Helper function to allocate memory at required address

init_rx_buf_pool
Initialize the receive buffer pool from where receive buffers are taken and returned

init_rx_queue
Steps to create and initialize a receive descriptor queue

init_tx_queue
Steps to create and initialize a transmit descriptor queue