Function nic_initialization::init_rx_buf_pool

pub fn init_rx_buf_pool(
    num_rx_buffers: usize, 
    buffer_size: u16, 
    rx_buffer_pool: &'static Queue<ReceiveBuffer>
) -> Result<(), &'static str>

Initialize the receive buffer pool from where receive buffers are taken and returned

Arguments

num_rx_buffers: number of buffers that are initially added to the pool

buffer_size: size of the receive buffers in bytes

rx_buffer_pool: buffer pool to initialize