Function nic_initialization::init_rx_queue

pub fn init_rx_queue<T: RxDescriptor>(
    num_desc: usize, 
    rx_buffer_pool: &'static Queue<ReceiveBuffer>, 
    buffer_size: usize, 
    rdbal: &mut Volatile<Rdbal>, 
    rdbah: &mut Volatile<Rdbah>, 
    rdlen: &mut Volatile<Rdlen>, 
    rdh: &mut Volatile<Rdh>, 
    rdt: &mut Volatile<Rdt>
) -> Result<(BoxRefMut<MappedPages, [T]>, Vec<ReceiveBuffer>), &'static str>

Steps to create and initialize a receive descriptor queue

Arguments

num_desc: number of descriptors in the queue

rx_buffer_pool: pool from which to take receive buffers

buffer_size: size of each buffer in the pool

rdbal: register to store the lower (least significant) 32 bits of the physical address of the array of receive descriptors

rdbah: register to store the higher (most significant) 32 bits of the physical address of the array of receive descriptors

rdlen: register to store the length in bytes of the array of receive descriptors

rdh: register to store the receive descriptor head index

rdt: register to store the receive descriptor tail index