Function nic_initialization::init_rx_queue

pub fn init_rx_queue<T: RxDescriptor, S: RxQueueRegisters>(
    num_desc: usize,
    rx_buffer_pool: &'static Queue<ReceiveBuffer>,
    buffer_size: usize,
    rxq_regs: &mut S
) -> Result<(BoxRefMut<MappedPages, [T]>, Vec<ReceiveBuffer>), &'static str>

Steps to create and initialize a receive descriptor queue

Arguments

num_desc: number of descriptors in the queue

rx_buffer_pool: pool from which to take receive buffers

buffer_size: size of each buffer in the pool in bytes

rxq_regs: registers needed to set up a receive queue